Проба фикса

+++ database/migrations/2021_10_02_020953_add_log_level_column.php
@@ -14,7 +14,7 @@ class AddLogLevelColumn extends Migration
     public function up()
     {
         Schema::table('logs', function (Blueprint $table) {
-            $table->smallInteger('level')->index();
+            $table->integer('level')->index();
         });
     }
